requirements.txt
face_recognition==1.3.0
Flask==3.0.0
google_api_python_client==2.108.0
joblib==1.2.0
numpy==1.26.2
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==10.1.0
protobuf==3.19.6
scikit_learn==1.2.2
gunicorn==21.2.0
